+++ src/tables/pes.h
@@ -3,7 +3,7 @@
 
 #define PES_PACKET_START_CODE_PREFIX 0x000001
 
-int pes_parse(const struct ts_header *header, const char *payload, uint8_t payload_len,
+int pes_parse(const struct ts_header *header, const char *payload, uint32_t payload_len,
 		struct demuxfs_data *priv);
 
 #endif /* __es_h */
