+++ src/tables/psi.c
@@ -80,7 +80,7 @@ static bool psi_check_header(struct psi_common_header *header)
 	return ret;
 }
 
-int psi_parse(struct psi_common_header *header, const char * payload, uint8_t payload_len)
+int psi_parse(struct psi_common_header *header, const char * payload, uint32_t payload_len)
 {
 	if (payload_len < 8) {
 		TS_WARNING("cannot parse PSI header: contents is smaller than 8 bytes (%d)", payload_len);
